Split out normalizePostForApi
