+++ client/state/posts/utils/index.js
@@ -1,15 +1,7 @@
 /**
  * External dependencies
  */
-import { flow } from 'lodash';
 import url from 'url';
-
-import { normalizeTermsForApi } from './normalize-terms-for-api';
-
-/**
- * Utility
- */
-const normalizeApiFlow = flow( [ normalizeTermsForApi ] );
 
 export { getNormalizedPostsQuery } from './get-normalized-posts-query';
 export { getSerializedPostsQuery } from './get-serialized-posts-query';
@@ -29,20 +21,7 @@ export { isAuthorEqual } from './is-author-equal';
 export { isDateEqual } from './is-date-equal';
 export { isStatusEqual } from './is-state-equal';
 export { getUnappliedMetadataEdits, areAllMetadataEditsApplied } from './metadata-edits';
-
-/**
- * Returns a normalized post object for sending to the API
- *
- * @param  {object} post Raw post object
- * @returns {object}      Normalized post object
- */
-export function normalizePostForApi( post ) {
-	if ( ! post ) {
-		return null;
-	}
-
-	return normalizeApiFlow( post );
-}
+export { normalizePostForApi } from './normalize-post-for-api';
 
 export const getEditURL = function( post, site ) {
 	if ( ! site ) {
